requirements.txt
annotated-doc==0.0.4
annotated-types==0.7.0
anyio==4.11.0
bcrypt==5.0.0
certifi==2025.11.12
cffi==2.0.0
click==8.3.1
colorama==0.4.6
cryptography==46.0.3
dnspython==2.8.0
email-validator==2.3.0
fastapi==0.122.0
fastapi-cli==0.0.16
fastapi-cloud-cli==0.5.2
fastar==0.8.0
h11==0.16.0
httpcore==1.0.9
httptools==0.7.1
httpx==0.28.1
idna==3.11
Jinja2==3.1.6
markdown-it-py==4.0.0
MarkupSafe==3.0.3
mdurl==0.1.2
mongoengine==0.29.1
pycparser==2.23
pydantic==2.12.5
pydantic_core==2.41.5
Pygments==2.19.2
PyJWT==2.10.1
pymongo==4.15.4
python-dotenv==1.2.1
python-multipart==0.0.20
PyYAML==6.0.3
rich==14.2.0
rich-toolkit==0.17.0
rignore==0.7.6
sentry-sdk==2.46.0
shellingham==1.5.4
sniffio==1.3.1
starlette==0.50.0
typer==0.20.0
typing-inspection==0.4.2
typing_extensions==4.15.0
urllib3==2.5.0
uvicorn==0.38.0
watchfiles==1.1.1
websockets==15.0.1
